requirements.txt
locustio
paramiko
ruamel.yaml